requirements.txt
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2
scikit-learn>=1.3.0
faker>=20.0.0
sentencepiece>=0.1.99
protobuf>=3.20.0
python-dotenv>=1.0.0
python-docx>=1.1.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
